+++ src/main/java/lagatrix/entities/dto/hardware/CPU.java
@@ -66,9 +66,9 @@ public class CPU implements Serializable {
 
     @Override
     public String toString() {
-        return String.format("Información de la CPU:\n\tModelo: %s\n\t"
-                + "Velocidad mínima: %s MHz\n\tVelocidad máxima: %s MHz\n\t"
-                + "Núcleos: %s\n\tHilos: %s\n\tMemoria Cahcé: %s", model, 
+        return String.format("\tInformación de la CPU:\n\t\tModelo: %s\n\t\t"
+                + "Velocidad mínima: %s MHz\n\t\tVelocidad máxima: %s MHz\n\t\t"
+                + "Núcleos: %s\n\t\tHilos: %s\n\t\tMemoria Cahcé: %s", model, 
                 minSpeed, maxSpeed, cores, threads, cacheMemory);
     }
 }